NavigationLink.vue: refactoring components
